requirements.txt
pytest==8.0.0
selenium==4.18.1
PyYAML==6.0.1
